# aspnetcore/src/IO.Swagger/Controllers/ProductsApi.cs
/*
 * Dilmah-Admin-BFF
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ProductsApiController : ControllerBase
    { 
        /// <summary>
        /// Get all channels of a group
        /// </summary>
        /// <param name="authToken">Authorization bearea token</param>
        /// <param name=""></param>
        /// <response code="200">success</response>
        [HttpGet]
        [Route("/channels/{channelGroup}")]
        [ValidateModelState]
        [SwaggerOperation("ChannelsChannelGroupGet")]
        public virtual IActionResult ChannelsChannelGroupGet([FromHeader][Required()]string authToken, [FromRoute][Required] )
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Delete a product
        /// </summary>
        /// <param name=""></param>
        /// <response code="200">success</response>
        [HttpDelete]
        [Route("/products/{code}")]
        [ValidateModelState]
        [SwaggerOperation("ProductsCodeDelete")]
        public virtual IActionResult ProductsCodeDelete([FromRoute][Required] )
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Get a product by Id
        /// </summary>
        /// <param name="authToken">Authorization bearea token</param>
        /// <param name=""></param>
        /// <response code="200">success</response>
        [HttpGet]
        [Route("/products/{code}")]
        [ValidateModelState]
        [SwaggerOperation("ProductsCodeGet")]
        public virtual IActionResult ProductsCodeGet([FromHeader][Required()]string authToken, [FromRoute][Required] )
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Get all products
        /// </summary>
        /// <param name="authToken">Authorization bearea token</param>
        /// <response code="200">success</response>
        [HttpGet]
        [Route("/products")]
        [ValidateModelState]
        [SwaggerOperation("ProductsGet")]
        public virtual IActionResult ProductsGet([FromHeader][Required()]string authToken)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Save product
        /// </summary>
        /// <response code="200">success</response>
        [HttpPost]
        [Route("/products")]
        [ValidateModelState]
        [SwaggerOperation("ProductsPost")]
        public virtual IActionResult ProductsPost()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }
    }
}