# aspnetcore/src/IO.Swagger/Controllers/TaxonApi.cs
/*
 * Dilmah-Admin-BFF
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class TaxonApiController : ControllerBase
    { 
        /// <summary>
        /// get all channel groups
        /// </summary>
        /// <param name="authToken">Authorization bearea token</param>
        /// <response code="200">success</response>
        [HttpGet]
        [Route("/channel-group")]
        [ValidateModelState]
        [SwaggerOperation("ChannelGroupGet")]
        public virtual IActionResult ChannelGroupGet([FromHeader][Required()]string authToken)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// get currency by group
        /// </summary>
        /// <param name="authToken">Authorization bearea token</param>
        /// <param name=""></param>
        /// <response code="200">success</response>
        [HttpGet]
        [Route("/currency/{channelGroup}")]
        [ValidateModelState]
        [SwaggerOperation("CurrencyChannelGroupGet")]
        public virtual IActionResult CurrencyChannelGroupGet([FromHeader][Required()]string authToken, [FromRoute][Required] )
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// get locale by group
        /// </summary>
        /// <param name="authToken">Authorization bearea token</param>
        /// <param name=""></param>
        /// <response code="200">success</response>
        [HttpGet]
        [Route("/locale/{channelGroup}")]
        [ValidateModelState]
        [SwaggerOperation("LocaleChannelGroupGet")]
        public virtual IActionResult LocaleChannelGroupGet([FromHeader][Required()]string authToken, [FromRoute][Required] )
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// get taxon list by group
        /// </summary>
        /// <param name="authToken">Authorization bearea token</param>
        /// <param name=""></param>
        /// <response code="200">success</response>
        [HttpGet]
        [Route("/taxon/{channelGroup}")]
        [ValidateModelState]
        [SwaggerOperation("TaxonChannelGroupGet")]
        public virtual IActionResult TaxonChannelGroupGet([FromHeader][Required()]string authToken, [FromRoute][Required] )
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// save taxon
        /// </summary>
        /// <response code="200">success</response>
        [HttpPost]
        [Route("/taxon")]
        [ValidateModelState]
        [SwaggerOperation("TaxonPost")]
        public virtual IActionResult TaxonPost()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }
    }
}