# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Dilmah-Admin-BFF",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/channel-group" : {
      "get" : {
        "tags" : [ "Taxon" ],
        "summary" : "get all channel groups",
        "parameters" : [ {
          "name" : "Auth-token",
          "in" : "header",
          "description" : "Authorization bearea token",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "data" : [ {
                        "id" : 1,
                        "code" : "DILMAH",
                        "name" : "Dilmah",
                        "description" : "Enjoye",
                        "is_enable" : 1
                      } ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/taxon/{channelGroup}" : {
      "get" : {
        "tags" : [ "Taxon" ],
        "summary" : "get taxon list by group",
        "parameters" : [ {
          "name" : "Auth-token",
          "in" : "header",
          "description" : "Authorization bearea token",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "channelGroup",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "date" : [ {
                        "id" : 9,
                        "code" : "black_tea",
                        "parent_id" : 24,
                        "level" : 1,
                        "position" : 1,
                        "name" : "Black Tea"
                      } ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/taxon" : {
      "post" : {
        "tags" : [ "Taxon" ],
        "summary" : "save taxon",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "body" : {
                  "value" : {
                    "code" : "toys",
                    "name" : "Toys",
                    "parentId" : 1,
                    "level" : 4,
                    "position" : 1,
                    "group" : "DILMAH",
                    "translations" : {
                      "en_US" : {
                        "name" : "Toys"
                      }
                    }
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "message" : "Successfully updated"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/locale/{channelGroup}" : {
      "get" : {
        "tags" : [ "Taxon" ],
        "summary" : "get locale by group",
        "parameters" : [ {
          "name" : "Auth-token",
          "in" : "header",
          "description" : "Authorization bearea token",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "channelGroup",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "data" : {
                        "id" : 3,
                        "code" : "en_Us_T-LOUNGE",
                        "group_code" : "T-LOUNGE",
                        "locale" : "en_Us"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/currency/{channelGroup}" : {
      "get" : {
        "tags" : [ "Taxon" ],
        "summary" : "get currency by group",
        "parameters" : [ {
          "name" : "Auth-token",
          "in" : "header",
          "description" : "Authorization bearea token",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "channelGroup",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "data" : {
                        "id" : 1,
                        "code" : "RS_T_LOUNGE",
                        "group_code" : "T-LOUNGE",
                        "currency" : "RS"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/products" : {
      "get" : {
        "tags" : [ "Products" ],
        "summary" : "Get all products",
        "parameters" : [ {
          "name" : "Auth-token",
          "in" : "header",
          "description" : "Authorization bearea token",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "data" : [ {
                        "code" : "black_tea_1",
                        "name" : "Premium Ceylon Black Tea - 25 Tea Bags",
                        "image" : "",
                        "main_taxon" : "Black Tea",
                        "enabled" : true
                      } ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Products" ],
        "summary" : "Save product",
        "requestBody" : {
          "description" : "body",
          "content" : {
            "application/json" : {
              "examples" : {
                "body" : {
                  "value" : {
                    "code" : "",
                    "enabled" : true,
                    "mainTaxon" : "mugs",
                    "productTaxons" : "mugs",
                    "options" : [ "mug_type" ],
                    "variant_selection_method" : "",
                    "channels" : [ "dilmah_1", "dilmah_2" ],
                    "translations" : {
                      "en_US" : {
                        "name" : "Theme Mug",
                        "slug" : "theme-mug"
                      }
                    },
                    "attributes" : [ {
                      "attribute" : "mug_color",
                      "localeCode" : "en_US",
                      "value" : "yellow"
                    } ],
                    "associations" : {
                      "similar_products" : "SMM,BMM"
                    }
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "message" : "successfull"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/products/{code}" : {
      "get" : {
        "tags" : [ "Products" ],
        "summary" : "Get a product by Id",
        "parameters" : [ {
          "name" : "Auth-token",
          "in" : "header",
          "description" : "Authorization bearea token",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "data" : {
                        "id" : "",
                        "code" : "black_tea_1",
                        "name" : "Premium Ceylon Black Tea - 25 Tea Bags",
                        "image" : "",
                        "main_taxon" : "Black Tea",
                        "enabled" : true,
                        "attributes" : [ {
                          "code" : "mug_material",
                          "name" : "Mug material",
                          "value" : "concrete",
                          "type" : "text",
                          "id" : 155
                        } ],
                        "options" : [ {
                          "id" : 1,
                          "code" : "mug_type",
                          "position" : 0,
                          "values" : [ {
                            "name" : "Mug type",
                            "code" : "mug_type_medium"
                          }, {
                            "name" : "Mug type",
                            "code" : "mug_type_double"
                          }, {
                            "name" : "Mug type",
                            "code" : "mug_type_monster"
                          } ]
                        } ],
                        "associations" : [ ],
                        "productTaxons" : [ ],
                        "channels" : [ {
                          "id" : 1,
                          "code" : "US_WEB",
                          "name" : "US Web Store",
                          "enabled" : true,
                          "taxCalculationStrategy" : "order_items_based"
                        } ],
                        "translations" : {
                          "en_US" : {
                            "locale" : "en_US",
                            "id" : 71,
                            "name" : "Theme Mug",
                            "slug" : "theme-mug"
                          },
                          "pl" : {
                            "locale" : "pl",
                            "id" : 72,
                            "name" : "Kubek z motywem",
                            "slug" : "kubek-z-motywem"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Products" ],
        "summary" : "Delete a product",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "message" : "successfull"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/channels/{channelGroup}" : {
      "get" : {
        "tags" : [ "Products" ],
        "summary" : "Get all channels of a group",
        "parameters" : [ {
          "name" : "Auth-token",
          "in" : "header",
          "description" : "Authorization bearea token",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "channelGroup",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "data" : [ {
                        "code" : "US_WEB",
                        "name" : "US Web Store",
                        "theme" : "",
                        "enabled" : true
                      } ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}